ThreadsX.jl

ThreadsX — Module

Threads⨉: Parallelized Base functions

tl;dr

Add prefix ThreadsX. to functions from Base to get some speedup, if supported. Example:

using ThreadsX
ThreadsX.sum(sin, 1:10_000)

To find out functions supported by ThreadsX.jl, just type ThreadsX. + <kbd>TAB</kbd> in the REPL:

julia> using ThreadsX

julia> ThreadsX.
MergeSort       any             findlast        maximum         sort!
QuickSort       count           foreach         minimum         sum
Set             extrema         map             prod            unique
StableQuickSort findall         map!            reduce
all             findfirst       mapreduce       sort

API

ThreadsX.jl is aiming at providing API compatible with Base functions to easily parallelize Julia programs.

All functions that exist directly under ThreadsX namespace are public API and they implement a subset of API provided by Base. Everything inside ThreadsX.Implementations is implementation detail. The public API functions of ThreadsX expect that the data structure and function(s) passed as argument are thread-safe. For example, ThreadsX.sum(f, array) assumes that executing f(::eltype(array)) and accessing elements as in array[i] from multiple threads is safe.

In addition to the Base API, all functions accept keyword argument basesize::Integer to configure the number of elements processed by each thread. A large value is useful for minimizing the overhead of using multiple threads. A small value is useful for load balancing when the time to process single item varies a lot from item to item. The default value of basesize for each function is currently an implementation detail.

Limitations

Keyword argument dims is not supported yet.

(There are probably more.)

Implementations

Most of reduce-based functions are implemented as a thin wrapper of Transducers.jl.

ThreadsX.sort! — Function

ThreadsX.sort!(xs; [smallsort, smallsize, basesize, alg, lt, by, rev, order])

Keyword Arguments

alg :: Base.Sort.Algorithm: ThreadsX.MergeSort, ThreadsX.QuickSort, ThreadsX.StableQuickSort etc. Base.MergeSort and Base.QuickSort can be used as aliases of ThreadsX.MergeSort and ThreadsX.QuickSort.

smallsort :: Union{Nothing,Base.Sort.Algorithm}: The algorithm to use for sorting small chunk of the input array.

smallsize :: Union{Nothing,Integer}: Size of array under which smallsort algorithm is used. nothing (default) means to use basesize.

basesize :: Union{Nothing,Integer}. Granularity of parallelization. nothing (default) means to choose the default size.

Other keyword arguments are passed to Base.sort!.

ThreadsX.sort — Function

ThreadsX.sort(xs; [smallsort, smallsize, basesize, alg, lt, by, rev, order])

See also ThreadsX.sort!.

ThreadsX.MergeSort — Constant

ThreadsX.MergeSort

Parallel merge sort algorithm.

Examples

ThreadsX.sort!(x; alg = MergeSort)

is more or less equivalent to

sort!(x; alg = ThreadsX.MergeSort)

although ThreadsX.sort! may be faster for very large integer arrays as it also parallelize counting sort.

ThreadsX.MergeSort is a Base.Sort.Algorithm, just like Base.MergeSort. It has a few properties for configuring the algorithm.

julia> using ThreadsX

julia> ThreadsX.MergeSort isa Base.Sort.Algorithm
true

julia> ThreadsX.MergeSort.smallsort === Base.Sort.DEFAULT_STABLE
true

The properties can be "set" by calling the algorithm object itself. A new algorithm object with new properties given by the keyword arguments is returned:

julia> alg = ThreadsX.MergeSort(smallsort = QuickSort) :: Base.Sort.Algorithm;

julia> alg.smallsort == QuickSort
true

julia> alg2 = alg(basesize = 64, smallsort = InsertionSort);

julia> alg2.basesize
64

julia> alg2.smallsort === InsertionSort
true

Properties

smallsort :: Base.Sort.Algorithm: Default to Base.Sort.DEFAULT_STABLE.

smallsize :: Union{Nothing,Integer}: Size of array under which smallsort algorithm is used. nothing (default) means to use basesize.

basesize :: Union{Nothing,Integer}. Base case size of parallel merge. nothing (default) means to choose the default size.