ThreadsX.jl

ThreadsX — Module

Threads⨉: Parallelized Base functions

tl;dr

Add prefix ThreadsX. to functions from Base to get some speedup, if supported. Example:

using ThreadsX
ThreadsX.sum(sin, 1:10_000)

To find out functions supported by ThreadsX.jl, just type ThreadsX. + TAB in the REPL:

julia> using ThreadsX

julia> ThreadsX.
MergeSort       any             findlast        maximum         sort!
QuickSort       count           foreach         minimum         sum
Set             extrema         map             prod            unique
StableQuickSort findall         map!            reduce
all             findfirst       mapreduce       sort

API

ThreadsX.jl is aiming at providing API compatible with Base functions to easily parallelize Julia programs.

All functions that exist directly under ThreadsX namespace are public API and they implement a subset of API provided by Base. Everything inside ThreadsX.Implementations is implementation detail. The public API functions of ThreadsX expect that the data structure and function(s) passed as argument are thread-safe. For example, ThreadsX.sum(f, array) assumes that executing f(::eltype(array)) and accessing elements as in array[i] from multiple threads is safe.

In addition to the Base API, all functions accept keyword argument basesize::Integer to configure the number of elements processed by each thread. A large value is useful for minimizing the overhead of using multiple threads. A small value is useful for load balancing when the time to process single item varies a lot from item to item. The default value of basesize for each function is currently an implementation detail.

Limitations

Keyword argument dims is not supported yet.

(There are probably more.)

Implementations

Most of reduce-based functions are implemented as thin wrappers of Transducers.jl.

Custom collections can support ThreadsX.jl API by implementing SplittablesBase.jl interface.

ThreadsX.foreach — Function

ThreadsX.foreach(f, arrays...; basesize)

A parallel version of

for args in zip(arrays...)
    f(args...)
end

ThreadsX.foreach uses linear and Cartesian indexing of arrays appropriately. However, it is likely very slow for sparse arrays.

Examples

julia> using ThreadsX

julia> xs = 1:10; ys = similar(xs);

julia> ThreadsX.foreach(eachindex(ys, xs)) do I
           @inbounds ys[I] = xs[I]
       end

As foreach can only be used for side-effects, it is likely that it has to be used with eachindex.

To avoid cumbersome indexing, a powerful pattern is to use Referenceables.jl with foreach:

julia> using Referenceables  # exports `referenceable`

julia> ThreadsX.foreach(referenceable(ys), xs) do y, x
           y[] = x
       end

Note that y[] does not have to be marked by @inbounds as it is ensured to be the reference to the valid location in the array.

Above function can also be written using map!. foreach is useful when, e.g., there are multiple outputs:

julia> A = randn(10, 10); sums = similar(A); muls = similar(A);

julia> ThreadsX.foreach(referenceable(sums), referenceable(muls), A, A') do s, m, x, y
           s[] = x + y
           m[] = x * y
       end

Above code fuses the computation of sums .= A .+ A' and muls .= A .* A' and runs it in parallel.

foreach can also be used when the array is both input and output:

julia> ThreadsX.foreach(referenceable(A)) do x
           x[] *= 2
       end

ThreadsX.map! — Function

ThreadsX.map!(f, dest, inputs...; basesize)

Parallelized map!.

ThreadsX.sort! — Function

ThreadsX.sort!(xs; [smallsort, smallsize, basesize, alg, lt, by, rev, order])

Sort a vector xs in parallel.

Examples

julia> using ThreadsX

julia> ThreadsX.sort!([9, 5, 2, 0, 1])
5-element Array{Int64,1}:
 0
 1
 2
 5
 9

julia> ThreadsX.sort!([0:5;]; alg = ThreadsX.StableQuickSort, by = _ -> 1)
6-element Array{Int64,1}:
 0
 1
 2
 3
 4
 5

It is also possible to use Base.sort! directly by specifying alg to be one of the parallel sort algorithms provided by ThreadsX:

julia> sort!([9, 5, 2, 0, 1]; alg = ThreadsX.MergeSort)
5-element Array{Int64,1}:
 0
 1
 2
 5
 9

This entry point may be slower than ThreadsX.sort! if the input is a very large array of integers with small range. In this case, ThreadsX.sort! uses parallel counting sort whereas sort! uses sequential counting sort.

Keyword Arguments

alg :: Base.Sort.Algorithm: ThreadsX.MergeSort, ThreadsX.QuickSort, ThreadsX.StableQuickSort etc. Base.MergeSort and Base.QuickSort can be used as aliases of ThreadsX.MergeSort and ThreadsX.QuickSort.

smallsort :: Union{Nothing,Base.Sort.Algorithm}: The algorithm to use for sorting small chunk of the input array.

smallsize :: Union{Nothing,Integer}: Size of array under which smallsort algorithm is used. nothing (default) means to use basesize.

basesize :: Union{Nothing,Integer}. Granularity of parallelization. nothing (default) means to choose the default size.

For keyword arguments, see Base.sort!.

ThreadsX.sort — Function

ThreadsX.sort(xs; [smallsort, smallsize, basesize, alg, lt, by, rev, order])

See also ThreadsX.sort!.

ThreadsX.MergeSort — Constant

ThreadsX.MergeSort

Parallel merge sort algorithm.

See also ThreadsX.QuickSort.

Examples

ThreadsX.MergeSort is a Base.Sort.Algorithm, just like Base.MergeSort. It has a few properties for configuring the algorithm.

julia> using ThreadsX

julia> ThreadsX.MergeSort isa Base.Sort.Algorithm
true

julia> ThreadsX.MergeSort.smallsort === Base.Sort.DEFAULT_STABLE
true

The properties can be "set" by calling the algorithm object itself. A new algorithm object with new properties given by the keyword arguments is returned:

julia> alg = ThreadsX.MergeSort(smallsort = QuickSort) :: Base.Sort.Algorithm;

julia> alg.smallsort == QuickSort
true

julia> alg2 = alg(basesize = 64, smallsort = InsertionSort);

julia> alg2.basesize
64

julia> alg2.smallsort === InsertionSort
true

Properties

smallsort :: Base.Sort.Algorithm: Default to Base.Sort.DEFAULT_STABLE.

smallsize :: Union{Nothing,Integer}: Size of array under which smallsort algorithm is used. nothing (default) means to use basesize.

basesize :: Union{Nothing,Integer}. Base case size of parallel merge. nothing (default) means to choose the default size.

ThreadsX.QuickSort — Constant

ThreadsX.QuickSort
ThreadsX.StableQuickSort

Parallel quick sort algorithms.

See also ThreadsX.MergeSort.

Properties

smallsort :: Base.Sort.Algorithm: Default to Base.Sort.DEFAULT_UNSTABLE.

smallsize :: Union{Nothing,Integer}: Size of array under which smallsort algorithm is used. nothing (default) means to use basesize.

basesize :: Union{Nothing,Integer}. Granularity of parallelization. nothing (default) means to choose the default size.